requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
cyminiball==2.1.1
Cython==0.29.35
einops==0.3.0
fonttools==4.47.2
idna==3.6
imageio==2.33.1
importlib-metadata
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
matplotlib==3.6.3
networkx==3.1
numba
numpy
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pybind11==2.11.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
tifffile==2023.7.10
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.1
trimesh==4.0.9
typing_extensions==4.9.0
urllib3==2.1.0
zipp
